refactor(bridge): carve /today /week /tasks handlers to telegram_handlers/ (#79 Phase I)
Ninth carve. Three Telegram cmd handlers + their shared arg parser
move to a new `telegram_handlers/` package. Behavior unchanged.

Symbols moved (4):
- _parse_by_flag — `/today by:model`, `by:profile` arg shorthand
- cmd_today, cmd_week, cmd_tasks

These are the most self-contained cmd handler group: their only
bot.py-internal dep was CHAT_ID, which the new module reads from the
same env var bot.py uses (TELEGRAM_CHAT_ID — single source of truth,
no cross-import). All other deps already live in task_agg/agg.py
since Phase D.

This also seeds the `telegram_handlers/` namespace for future
phases (cmd_budget, cmd_pricing, cmd_usage, cmd_chats, etc).
The other cmd handler groups still hold bot.py-internal globals
(monitor_*, az_context, send_telegram), so they need carving in
order: monitor state → send_telegram → those handler groups.

Tests added (6):
- _parse_by_flag default (no args), model variants, profile variants,
  case-insensitive, unknown returns None
- CHAT_ID read from env confirmed

Async cmd handlers themselves aren't unit-tested — they need a
python-telegram-bot test client which is heavier than this PR's
worth. The pure helper is the regression risk worth pinning.

Live verified after rebuild:
- bridge container starts clean
- 128/128 pytest pass (122 prior + 6 new)
- ruff clean

bot.py: 2,039 → 1,839 lines (-200). Total since start: 3,579 → 1,839
(-1,740, -49%).
