+++ telegram-bridge/telegram_handlers/__init__.py
@@ -1,0 +1,12 @@
+"""Telegram command handlers — Phase I+ carve from bot.py (issue #79).
+
+The /today, /week, /tasks trio gets carved here first because they're
+the most self-contained group: they only depend on task_agg + the
+chat-ID env var. Future phases can carve more cmd handlers (budget,
+pricing, usage, system, monitor toggles, …) into their own files
+under this package as their own bot.py-internal deps come unstuck.
+"""
+
+from .today import _parse_by_flag, cmd_tasks, cmd_today, cmd_week
+
+__all__ = ["_parse_by_flag", "cmd_tasks", "cmd_today", "cmd_week"]
